Fix missing textures cbmultipart (#6)
Some of the CB Multipart mod elements are not rendering properly with Magnesium. Remove Magnesium, and add Rubidium, which is the maintained, and supported version, and like Magnesium, it is a fork of the very same SODIUM mod. Also updates CB Multipart, and its dependency CodeChickenLib to the latest available version